CRF Added new identifier superclass
